backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.4.0
pymongo==4.6.2
python-dotenv==1.0.1
pydantic==2.6.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
starlette==0.36.3
aiohttp==3.9.3
spacy==3.7.4
python-multipart==0.0.9